refactor: layout classes packages
