+++ video-sdk/src/main/java/com/kaleyra/video_sdk/call/stream/layoutsystem/itemsprovider/MosaicStreamItemsProvider.kt
@@ -1,4 +1,4 @@
-package com.kaleyra.video_sdk.call.stream.viewmodel
+package com.kaleyra.video_sdk.call.stream.layoutsystem.itemsprovider
 
 import com.kaleyra.video_sdk.call.stream.model.StreamItem
 import com.kaleyra.video_sdk.call.stream.model.MoreStreamsUserPreview
