+++ video-sdk/src/test/java/com/kaleyra/video_sdk/layout/FeaturedStreamItemsProviderTest.kt
@@ -3,7 +3,7 @@ package com.kaleyra.video_sdk.layout
 import com.kaleyra.video_sdk.call.stream.model.StreamItem
 import com.kaleyra.video_sdk.call.stream.model.MoreStreamsUserPreview
 import com.kaleyra.video_sdk.call.stream.model.core.StreamUi
-import com.kaleyra.video_sdk.call.stream.viewmodel.FeaturedStreamItemsProviderImpl
+import com.kaleyra.video_sdk.call.stream.layoutsystem.itemsprovider.FeaturedStreamItemsProviderImpl
 import com.kaleyra.video_sdk.call.stream.model.StreamItemState
 import com.kaleyra.video_sdk.common.avatar.model.ImmutableUri
 import io.mockk.mockk
