+++ video-sdk/src/test/java/com/kaleyra/video_sdk/layout/ManualLayoutTest.kt
@@ -5,10 +5,10 @@ package com.kaleyra.video_sdk.layout
 import com.kaleyra.video_sdk.call.stream.model.StreamItem
 import com.kaleyra.video_sdk.call.stream.model.StreamItemState
 import com.kaleyra.video_sdk.call.stream.model.core.StreamUi
-import com.kaleyra.video_sdk.call.stream.viewmodel.FeaturedStreamItemsProvider
-import com.kaleyra.video_sdk.call.stream.viewmodel.ManualLayoutImpl
-import com.kaleyra.video_sdk.call.stream.viewmodel.MosaicStreamItemsProvider
-import com.kaleyra.video_sdk.call.stream.viewmodel.StreamLayoutConstraints
+import com.kaleyra.video_sdk.call.stream.layoutsystem.itemsprovider.FeaturedStreamItemsProvider
+import com.kaleyra.video_sdk.call.stream.layoutsystem.layout.ManualLayoutImpl
+import com.kaleyra.video_sdk.call.stream.layoutsystem.itemsprovider.MosaicStreamItemsProvider
+import com.kaleyra.video_sdk.call.stream.layoutsystem.config.StreamLayoutConstraints
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
